Moved class around
